modified:   .gitignore 	new file:   backend/press/alg_exec.go 	new file:   backend/press/alg_gzip.go 	new file:   backend/press/alg_lz4.go 	new file:   backend/press/alg_snappy.go 	new file:   backend/press/benchmarks.txt 	modified:   backend/press/compression.go 	modified:   backend/press/compression_test.go 	new file:   backend/press/notes.txt 	modified:   go.mod 	modified:   go.sum
