+++ .gitignore
@@ -5,3 +5,4 @@ build
 docs/public
 rclone.iml
 .idea
+backend/press/test.vdi
