+++ backend/press/notes.txt
@@ -1,0 +1,26 @@
+Notes:
+Structure of our metadata at the end is:
+gzipExtraify(gzip([4-byte header size][4-byte block size] ... [4-byte block size][4-byte raw size of last block]))
+
+LZ4 implementation:
+Flags:
+	Version = 01
+	Independent = 1
+	Block Checksum = 1
+	Content Size = 0
+	Content Checksum = 0
+	Reserved = 0
+	Dictionary ID = 0
+
+BD byte:
+	Reserved = 0
+	Block Max Size = 101 (or 5; 256kb)
+	Reserved = 0000
+
+[Header Checksum byte]
+
+[Blocks with xxhash checksums]
+
+[4 null bytes]
+
+[gzip stuff as trailing garbage]
